Fixed std::unordered_map work on VFSEntriesMap creation.
Added elements rehashed automatically if the new number of elements is greater than
max_load_factor()*bucket_count(). In the same time, bucket_count will be doubled, as soon,
as limit are reached (but not increased by 1). This mean, we could have a bunch of created
elements at the end of map creation, that are not rehash the map.
http://en.cppreference.com/w/cpp/container/unordered_map/operator_at