Fixed mess with array parameters in functions and methods.
Previously, we provide fixed size array as pointer, instead of reference to array
with fixed size. In this case, compiler would not care what exactly we are sending to
functions and methods as parameter. As an example of such issue in code:
https://github.com/viewizard/astromenace/blob/8c1a75428fbc2bf537c25ebe3a59c28c0773a601/src/menu/menu_information.cpp#L1864
tmp_matrix[33] should be tmp_matrix[9] for sure (some kind of misspelling for matrix 3x3?).
Cosmetic changes.