Fixed segfault on mission start after game restart.
In case we change options and restart the game from the mission,
object manager will delete PlayerFighter object during restart.
This mean, we have already released memory address in PlayerFighter.
PlayerFighter should be released in RealExitGame(),
we don't really need this check in InitGamePlayerShip().
This is why we need smart pointers...