Revised menu code.
Version and copyright text rendering moved from internal
size to real, no reason in huge fonts for this information
for large screen resolution.
Renamed GAME_BUILD to GAME_VFS_BUILD, now will not rendered
in menu. Previously, this build number was planned as vfs only in
order to prevent usage not appropriate gamedata file, so, we back
to the initial usage.