Add project-wide namespace "viewizard".
We don't use "astromenace" name here, since this name will be used for game code
in order to provide strict interaction with "core" code.
"core" code should not interact with "game" code itself in any way (since
it may be used in other games), so, it will be braced by own "core" namespace.